Add resume option to Forward

+++ DataLoader.lua
@@ -40,20 +40,24 @@ function DataLoader:__init(...)
 
   -- find the image paths
   print('Finding images...')
-  local imageList = os.tmpname()
-  local tmpfile = os.tmpname()
-  local tmphandle = assert(io.open(tmpfile, 'w'))
-  local command = 'find ' .. args.path .. ' ' .. findOptions .. ' >>"' .. imageList .. '" \n'
-  tmphandle:write(command)
-  io.close(tmphandle)
-  os.execute('bash ' .. tmpfile)
-  os.execute('rm -f ' .. tmpfile)
+  if paths.filep(args.path) then
+    self.imageList = args.path
+  else
+    self.imageList = os.tmpname()
+    local tmpfile = os.tmpname()
+    local tmphandle = assert(io.open(tmpfile, 'w'))
+    local command = 'find ' .. args.path .. ' ' .. findOptions .. ' >>"' .. self.imageList .. '" \n'
+    tmphandle:write(command)
+    io.close(tmphandle)
+    os.execute('bash ' .. tmpfile)
+    os.execute('rm -f ' .. tmpfile)
 
-  --==========================================================================
-  local length = tonumber(sys.fexecute('wc -l "' .. imageList .. '" | ' .. 'cut -f1 -d" "'))
-  assert(length > 0, 'Could not find any image file in the given input paths')
+    --==========================================================================
+    local length = tonumber(sys.fexecute('wc -l "' .. self.imageList .. '" | ' .. 'cut -f1 -d" "'))
+    assert(length > 0, 'Could not find any image file in the given input paths')
+  end
 
-  self.imagePaths = io.open(imageList)
+  self.imagePaths = io.open(self.imageList)
   self.lineOffset = {0}
   for line in self.imagePaths:lines() do
     self.lineOffset[#self.lineOffset+1] = self.imagePaths:seek()
@@ -129,7 +133,7 @@ function DataLoader.loadInputs(pathNames, preprocessFn, workerFn)
   end
 end
 
-function DataLoader:runAsync(batchSize, epochSize, shuffle, preprocessFn, workerFn, resultHandler)
+function DataLoader:runAsync(batchSize, epochSize, shuffle, preprocessFn, workerFn, resultHandler, startIdx)
   if batchSize == -1 then
     batchSize = self:size()
   end
@@ -144,12 +148,24 @@ function DataLoader:runAsync(batchSize, epochSize, shuffle, preprocessFn, worker
     perm = torch.randperm(self:size())
   end
 
-  startJobs(epochSize)
-  for i=1,epochSize do
-    local indexStart = (i-1) * batchSize + 1
-    local indexEnd = (indexStart + batchSize - 1)
+  startIdx = startIdx or 1
+  if startIdx > epochSize then return end
+  startJobs(epochSize-startIdx+1)
+
+  local indexStart = (startIdx-1) * batchSize + 1
+  local indexEnd = min(indexStart + batchSize - 1, self:size())
+  for i=startIdx,epochSize do
+    local nextStart = indexEnd + 1
+    local nextEnd = min(nextStart + batchSize - 1, self:size())
+    if indexEnd < self:size() and nextEnd == nextStart then
+      -- next batch will be size 1, which we want to avoid
+      indexEnd = indexEnd - 1
+      nextStart = nextStart - 1
+    end
     local pathNames = self:get(indexStart, indexEnd, perm)
     threads:addjob(self.loadInputs, resultHandler, pathNames, preprocessFn, workerFn)
+    indexStart = nextStart
+    indexEnd = nextEnd
   end
   threads:synchronize()
 end
